+++ infrastructure/cdk/stacks/dynamodb_stack.py
@@ -100,6 +100,7 @@ class DynamoDBStack(Stack):
             ),
             billing_mode=dynamodb.BillingMode.PAY_PER_REQUEST,
             stream=dynamodb.StreamViewType.NEW_AND_OLD_IMAGES,
+            point_in_time_recovery=True,
             removal_policy=RemovalPolicy.RETAIN,
         )
 
@@ -147,6 +148,7 @@ class DynamoDBStack(Stack):
                 type=dynamodb.AttributeType.STRING,
             ),
             billing_mode=dynamodb.BillingMode.PAY_PER_REQUEST,
+            point_in_time_recovery=True,
             removal_policy=RemovalPolicy.RETAIN,
         )
 
@@ -196,6 +198,7 @@ class DynamoDBStack(Stack):
                 type=dynamodb.AttributeType.STRING,
             ),
             billing_mode=dynamodb.BillingMode.PAY_PER_REQUEST,
+            point_in_time_recovery=True,
             removal_policy=RemovalPolicy.RETAIN,
         )
 
@@ -241,6 +244,7 @@ class DynamoDBStack(Stack):
                 type=dynamodb.AttributeType.STRING,
             ),
             billing_mode=dynamodb.BillingMode.PAY_PER_REQUEST,
+            point_in_time_recovery=True,
             removal_policy=RemovalPolicy.RETAIN,
         )
 
@@ -272,6 +276,7 @@ class DynamoDBStack(Stack):
                 type=dynamodb.AttributeType.STRING,
             ),
             billing_mode=dynamodb.BillingMode.PAY_PER_REQUEST,
+            point_in_time_recovery=True,
             removal_policy=RemovalPolicy.RETAIN,
         )
 
@@ -317,6 +322,7 @@ class DynamoDBStack(Stack):
                 type=dynamodb.AttributeType.STRING,
             ),
             billing_mode=dynamodb.BillingMode.PAY_PER_REQUEST,
+            point_in_time_recovery=True,
             removal_policy=RemovalPolicy.RETAIN,
         )
 
@@ -376,6 +382,7 @@ class DynamoDBStack(Stack):
                 type=dynamodb.AttributeType.STRING,
             ),
             billing_mode=dynamodb.BillingMode.PAY_PER_REQUEST,
+            point_in_time_recovery=True,
             removal_policy=RemovalPolicy.RETAIN,
         )
 
@@ -407,6 +414,7 @@ class DynamoDBStack(Stack):
                 type=dynamodb.AttributeType.STRING,
             ),
             billing_mode=dynamodb.BillingMode.PAY_PER_REQUEST,
+            point_in_time_recovery=True,
             removal_policy=RemovalPolicy.RETAIN,
         )
 
@@ -452,6 +460,7 @@ class DynamoDBStack(Stack):
                 type=dynamodb.AttributeType.STRING,
             ),
             billing_mode=dynamodb.BillingMode.PAY_PER_REQUEST,
+            point_in_time_recovery=True,
             removal_policy=RemovalPolicy.RETAIN,
         )
 
@@ -525,6 +534,7 @@ class DynamoDBStack(Stack):
                 type=dynamodb.AttributeType.STRING,
             ),
             billing_mode=dynamodb.BillingMode.PAY_PER_REQUEST,
+            point_in_time_recovery=True,
             removal_policy=RemovalPolicy.RETAIN,
         )
 
@@ -598,6 +608,7 @@ class DynamoDBStack(Stack):
                 type=dynamodb.AttributeType.STRING,
             ),
             billing_mode=dynamodb.BillingMode.PAY_PER_REQUEST,
+            point_in_time_recovery=True,
             removal_policy=RemovalPolicy.RETAIN,
         )